Decompose scene interaction runtime
